+++ lnmp-docker.ps1
@@ -93,7 +93,7 @@ Function env_status(){
     Copy-Item secrets/minio/secret.example.txt secrets/minio/secret.txt
   }
 
-  _cp_only_not_exists docker-compose.include.example.yml docker-compose.include.yml
+  _cp_only_not_exists docker-lnmp.include.example.yml docker-lnmp.include.yml
 
   _cp_only_not_exists config/php/docker-php.ini.example config/php/docker-php.ini
   _cp_only_not_exists config/php/php.development.ini config/php/php.ini
@@ -418,7 +418,7 @@ Function get_compose_options($compose_files,$isBuild=0){
     $options +=" -f lnmp-include\$item\docker-compose.yml -f lnmp-include\$item\docker-compose.override.yml "
   }
 
-  $options += " -f docker-compose.include.yml "
+  $options += " -f docker-lnmp.include.yml "
 
   return $options
 }
@@ -459,32 +459,32 @@ switch($first){
     }
 
     build {
-      $options=get_compose_options "docker-compose.yml", `
-                                   "docker-compose.build.yml" `
+      $options=get_compose_options "docker-lnmp.yml", `
+                                   "docker-lnmp.build.yml" `
                                     1
 
       powershell -c "docker-compose $options build $other --parallel"
     }
 
     build-config {
-      $options=get_compose_options "docker-compose.yml", `
-                                   "docker-compose.build.yml" `
+      $options=get_compose_options "docker-lnmp.yml", `
+                                   "docker-lnmp.build.yml" `
                                     1
 
       powershell -c "docker-compose $options config $other"
     }
 
     build-up {
-      $options=get_compose_options "docker-compose.yml", `
-                                   "docker-compose.build.yml" `
+      $options=get_compose_options "docker-lnmp.yml", `
+                                   "docker-lnmp.build.yml" `
                                     1
 
       powershell -c "docker-compose $options up -d $other"
     }
 
     build-push {
-      $options=get_compose_options "docker-compose.yml", `
-                                   "docker-compose.build.yml" `
+      $options=get_compose_options "docker-lnmp.yml", `
+                                   "docker-lnmp.build.yml" `
                                     1
 
       powershell -c "docker-compose $options build $other --parallel"
@@ -499,8 +499,8 @@ switch($first){
     }
 
     config {
-      $options=get_compose_options "docker-compose.yml", `
-                                   "docker-compose.override.yml"
+      $options=get_compose_options "docker-lnmp.yml", `
+                                   "docker-lnmp.override.yml"
 
       powershell -c "docker-compose $options config $other"
     }
@@ -525,8 +525,8 @@ switch($first){
         $command = ${LNMP_SERVICES}
       }
 
-      $options=get_compose_options "docker-compose.yml", `
-                                   "docker-compose.override.yml"
+      $options=get_compose_options "docker-lnmp.yml", `
+                                   "docker-lnmp.override.yml"
 
       powershell -c "docker-compose $options up -d $command"
 
@@ -535,8 +535,8 @@ switch($first){
     }
 
     pull {
-      $options=get_compose_options "docker-compose.yml",`
-                                   "docker-compose.override.yml"
+      $options=get_compose_options "docker-lnmp.yml",`
+                                   "docker-lnmp.override.yml"
 
       powershell -c "docker-compose $options pull ${LNMP_SERVICES}"
 
@@ -545,7 +545,10 @@ switch($first){
     }
 
     down {
-      docker-compose down --remove-orphans
+      $options=get_compose_options "docker-lnmp.yml",`
+                                   "docker-lnmp.override.yml"
+
+      powershell -c "docker-compose $options down --remove-orphans"
 
       #@custom
       __lnmp_custom_down
@@ -589,8 +592,8 @@ switch($first){
     }
 
     restart {
-      $options=get_compose_options "docker-compose.yml", `
-                          "docker-compose.override.yml"
+      $options=get_compose_options "docker-lnmp.yml", `
+                          "docker-lnmp.override.yml"
 
       powershell -c "docker-compose $options restart $other"
       #@custom
@@ -718,8 +721,8 @@ switch($first){
     }
 
     clusterkit {
-       $options=get_compose_options "docker-compose.yml", `
-                                    "docker-compose.override.yml", `
+       $options=get_compose_options "docker-lnmp.yml", `
+                                    "docker-lnmp.override.yml", `
                                     "cluster/docker-cluster.mysql.yml", `
                                     "cluster/docker-cluster.redis.yml"
 
@@ -971,8 +974,8 @@ XXX
       cp pcit/.env.development ${APP_ROOT}/.pcit/.env.development
 
       # 启动
-      $options=get_compose_options "docker-compose.yml", `
-                                   "docker-compose.override.yml"
+      $options=get_compose_options "docker-lnmp.yml", `
+                                   "docker-lnmp.override.yml"
 
       & {docker-compose $options up -d ${LNMP_SERVICES} pcit}
     }
@@ -1019,6 +1022,7 @@ XXX
         -p 2376:2375 `
         --mount type=bind,src=/var/run/docker.sock,target=/var/run/docker.sock `
         -e PORT=2375 `
+        --health-cmd="wget 127.0.0.1:2375/info -O /proc/self/fd/2" `
         shipyard/docker-proxy
     }
 
@@ -1104,8 +1108,8 @@ This local server support Docker Desktop EDGE v2.0.1.0(30090)
         __lnmp_custom_command $args
         printInfo `
 "Exec docker-compose command, maybe you input command is notdefined, then output docker-compose help information"
-        $options=get_compose_options "docker-compose.yml",`
-                                     "docker-compose.override.yml"
+        $options=get_compose_options "docker-lnmp.yml",`
+                                     "docker-lnmp.override.yml"
 
         powershell -c "docker-compose $options $args"
       }